web/requirements.txt
asgiref==3.4.1
black==21.8b0
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
dj-database-url==0.5.0
Django==3.2.7
django-extensions==3.1.3
flake8==3.9.2
gunicorn==20.1.0
idna==3.2
mccabe==0.6.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.3.0
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pyflakes==2.3.1
pytz==2021.1
PyYAML==5.4.1
regex==2021.8.28
requests==2.26.0
sqlparse==0.4.2
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.6
whitenoise==5.3.0
